requirements.txt
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
backports.zoneinfo==0.2.1
coverage==6.2
Django==4.0.1
django-crispy-forms==1.13.0
django-filter==21.1
django-generic-filters==2.1.0
factory-boy==3.2.1
Faker==11.3.0
faker-vehicle==0.2.0
flake8==4.0.1
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
logilab-astng==0.24.3
logilab-common==1.8.3
mccabe==0.6.1
munch==2.5.0
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
wrapt==1.13.3
yapf==0.32.0
